+++ openopc2/__main__.py
@@ -1,0 +1,6 @@
+from openopc2.cli import OpcCli
+
+if __name__ == '__main__':
+    opc_cli = OpcCli()
+    print("*" * 100)
+    opc_cli.parse_arguments()
