+++ openopc2/gateway_server.py
@@ -12,7 +12,7 @@ import time
 import Pyro5.server
 
 from openopc2.config import open_opc_config
-from openopc2.opc_da_client import OpcDaClient, __version__
+from openopc2.da_client import OpcDaClient, __version__
 
 
 @Pyro5.api.expose
