+++ tests/test_config.py
@@ -1,16 +1,17 @@
 from unittest import TestCase
 
-from openopc2.config import open_opc_config
+from openopc2.config import OpenOpcConfig
+
 
 
 class TestConfig(TestCase):
 
     def test_print_config(self):
-        config = open_opc_config
+        config = OpenOpcConfig()
         config.print_config()
 
     def test_overwrite_config_runtime(self):
-        config = open_opc_config
+        config = OpenOpcConfig()
         self.assertEqual(config.OPC_CLIENT, 'OpenOPC')
         config.OPC_CLIENT = 'Shubi'
         self.assertEqual(config.OPC_CLIENT, 'Shubi')
